requirements.txt
anyio==3.6.1
attrs==22.1.0
autocommand==2.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
cheroot==8.6.0
CherryPy==18.8.0
decorator==5.1.1
dnspython==2.2.1
email-validator==1.3.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
inflect==6.0.0
iniconfig==1.1.1
ipaddress==1.0.23
jaraco.classes==3.2.3
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.9.1
loguru==0.6.0
more-itertools==8.14.0
packaging==21.3
phonenumbers==8.12.56
pluggy==1.0.0
portend==3.1.0
priority==2.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.10.2
pyparsing==3.0.9
pyQiwiP2P==2.0.6
pyTelegramBotAPI==4.7.0
pytest==7.1.3
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
response-report==1.1.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
tempora==5.0.2
toml==0.10.2
tomli==2.0.1
typing==3.7.4.3
typing_extensions==4.3.0
urllib3==1.26.12
validators==0.20.0
wsproto==1.2.0
zc.lockfile==2.0
